@@ -2,19 +2,14 @@
 
 namespace App\Command;
 
-use App\Service\System as SystemService;
 use App\Service\Bot as BotService;
-use Doctrine\ORM\EntityManagerInterface;
+use App\Service\System as SystemService;
 use Psr\Log\LoggerInterface;
-use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Input\ArrayInput;
-use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Output\NullOutput;
+use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
-use Symfony\Component\HttpKernel\KernelInterface;
 
 class ActivateBot extends Command
 {
@@ -55,22 +50,24 @@ class ActivateBot extends Command
     {
         $this->output = $output;
 
-        if ($this->allow_bots != "true") {
+        if ('true' != $this->allow_bots) {
             $this->output->writeln([
                 '<comment>Bots are not allowed to run.</comment>',
             ]);
+
             return 0;
         }
-        if ($input->getArgument('bot_id') != "all") {
+        if ('all' != $input->getArgument('bot_id')) {
             $this->runBot($input->getArgument('bot_id'));
         } else {
             // activate all bots
             foreach (scandir($this->params->get('dir.bots')) as $file) {
-                if ($file != "." && $file != ".." && is_dir($this->params->get('dir.bots')."/$file")) {
+                if ('.' != $file && '..' != $file && is_dir($this->params->get('dir.bots')."/$file")) {
                     $this->runBot($file);
                 }
             }
         }
+
         return 0;
     }
 
@@ -78,16 +75,16 @@ class ActivateBot extends Command
     {
         $this->system->set('last_bot_timestamp', time());
         $this->system->set('last_bot_id', $bot_id);
-        $this->system->set("bot::" . $bot_id, time());
-        $action_file = $this->botDir . "/" . $bot_id . "/action.php";
+        $this->system->set('bot::'.$bot_id, time());
+        $action_file = $this->botDir.'/'.$bot_id.'/action.php';
         if (file_exists($action_file) && is_readable($action_file)) {
             try {
                 set_error_handler(function ($errno, $errstr, $errfile, $errline) {
                     throw new \ErrorException($errstr, $errno, 0, $errfile, $errline);
                 });
-                include($action_file);
+                include $action_file;
             } catch (\Exception $e) {
-                $this->logger->error($bot_id.": ".$e->getMessage());
+                $this->logger->error($bot_id.': '.$e->getMessage());
             }
         } else {
             $this->output->writeln([
