@@ -2,9 +2,7 @@
 
 namespace App\Command;
 
-use App\Service\Bot as BotService;
 use App\Service\System;
-use Doctrine\ORM\EntityManagerInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Bundle\FrameworkBundle\Console\Application;
 use Symfony\Component\Console\Command\Command;
@@ -70,7 +68,7 @@ class Cron extends Command
                 'period' => intval($this->params->get('cron.bot.activate')),
                 'type' => 'light',
                 'options' => [
-                    'bot_id' => 'all'
+                    'bot_id' => 'all',
                 ],
             ],
             [
@@ -116,6 +114,7 @@ class Cron extends Command
         $this->input = $input;
         $this->output = $output;
         $this->runTask();
+
         return 0;
     }
 
@@ -129,6 +128,7 @@ class Cron extends Command
         // don't run if on a POST or an upload files request. Don't run twice in the same process
         if (!empty($_POST) || !empty($_FILES) || defined('TASK_RUNNING') || $this->running) {
             $this->logger->info('Task already running');
+
             return false;
         }
 
@@ -143,6 +143,7 @@ class Cron extends Command
         // if a task is running and it's not older than MAX_TASK_LOCK_DURATION, do nothing
         if ($task_running && isset($last_task_timestamp) && $last_task_timestamp > time() - intval($this->params->get('max_task_lock_duration'))) {
             $this->logger->notice($last_task_name.' is already running since '.(time() - $last_task_timestamp).'s');
+
             return false;
         }
 
@@ -167,7 +168,7 @@ class Cron extends Command
             // if it's a heavy task and we're not in the idle hours, don't do it
             if (
                 isset($task['type'])
-                && $task['type'] == 'heavy'
+                && 'heavy' == $task['type']
                 && (
                     (new \DateTime())->format('H') < $idle_hours[0]
                     || (new \DateTime())->format('H') > $idle_hours[1]
@@ -189,6 +190,7 @@ class Cron extends Command
         }
 
         $this->system->set('task_running', false);
+
         return true;
     }
 
@@ -199,7 +201,7 @@ class Cron extends Command
             $command = (new Application($this->kernel))->find($id);
             $returnCode = $command->run(new ArrayInput($options), $this->output ?? new NullOutput());
         } catch (\Exception $e) {
-            $this->logger->error($id . ' ' . $e->getMessage());
+            $this->logger->error($id.' '.$e->getMessage());
         } finally {
             if (isset($returnCode) && 0 != $returnCode) {
                 $this->logger->error("$id failed, return code: $returnCode");
