@@ -6,7 +6,6 @@ use App\Controller\ApiController;
 use App\Entity\Bookmark;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
-use Nelmio\ApiDocBundle\Annotation\Model;
 use Nelmio\ApiDocBundle\Annotation\Security;
 use OpenApi\Annotations as OA;
 use Symfony\Component\ExpressionLanguage\Expression;
@@ -27,11 +26,14 @@ class Delete extends ApiController
 
     /**
      * @Route("/bookmarks/{id}", methods={"DELETE"})
+     *
      * @OA\Response(
      *  response=204,
      *  description="Delete a bookmark"
      * )
+     *
      * @OA\Tag(name="bookmark")
+     *
      * @Security(name="api_key")
      */
     public function index(
