@@ -6,7 +6,6 @@ use App\Controller\ApiController;
 use App\Entity\File;
 use App\Service\Image as ImageService;
 use App\Service\Uuid;
-use Nelmio\ApiDocBundle\Annotation\Model;
 use Nelmio\ApiDocBundle\Annotation\Security;
 use OpenApi\Annotations as OA;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
@@ -19,14 +18,18 @@ class GetMiniature extends ApiController
 {
     /**
      * @Route("/images/{type}/{width}/{height}/{id}", methods={"GET"})
+     *
      * @OA\Response(
      *  response=200,
      *  description="Get a miniature of a file",
+     *
      *  @OA\Schema(
      *    type="binary",
      *  )
      * )
+     *
      * @OA\Tag(name="file")
+     *
      * @Security(name="api_key")
      */
     #[Cache(public: true, maxage: 31536000)]
