@@ -6,14 +6,14 @@ use App\Controller\ApiController;
 use App\Entity\Group;
 use App\Entity\Message;
 use Doctrine\ORM\EntityManagerInterface;
+use Nelmio\ApiDocBundle\Annotation\Model;
+use Nelmio\ApiDocBundle\Annotation\Security;
+use OpenApi\Annotations as OA;
 use Symfony\Component\ExpressionLanguage\Expression;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\SerializerInterface;
-use Nelmio\ApiDocBundle\Annotation\Model;
-use Nelmio\ApiDocBundle\Annotation\Security;
-use OpenApi\Annotations as OA;
 use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Contracts\Cache\TagAwareCacheInterface;
 
@@ -32,12 +32,16 @@ class Get extends ApiController
 
     /**
      * @Route("/groups/{id}", methods={"GET"})
+     *
      * @OA\Response(
      *  response=200,
      *  description="Get a group",
+     *
      *  @Model(type=App\Entity\Group::class, groups={"read_group"})
      * )
+     *
      * @OA\Tag(name="group")
+     *
      * @Security(name="api_key")
      */
     public function index(string $id): Response
@@ -54,7 +58,7 @@ class Get extends ApiController
         $cacheKey = 'group_'.$group->getId().'_info';
 
         $data = $this->cache->get($cacheKey, function (ItemInterface $item) use ($group) {
-            $item->expiresAfter(3600*24*7);
+            $item->expiresAfter(3600 * 24 * 7);
             $item->tag('group_'.$group->getId());
 
             return $this->serialize($group, ['read_group']);
@@ -68,12 +72,16 @@ class Get extends ApiController
 
     /**
      * @Route("/groups/{id}/random", methods={"GET"})
+     *
      * @OA\Response(
      *  response=200,
      *  description="Get a random message from the group",
+     *
      *  @Model(type=App\Entity\Message::class, groups={"read_message"})
      * )
+     *
      * @OA\Tag(name="group")
+     *
      * @Security(name="api_key")
      */
     public function random_message(string $id): Response
