@@ -7,7 +7,6 @@ use App\Entity\Message;
 use App\Entity\Notification;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
-use Nelmio\ApiDocBundle\Annotation\Model;
 use Nelmio\ApiDocBundle\Annotation\Security;
 use OpenApi\Annotations as OA;
 use Symfony\Component\ExpressionLanguage\Expression;
@@ -33,11 +32,14 @@ class Delete extends ApiController
 
     /**
      * @Route("/messages/{id}", methods={"DELETE"})
+     *
      * @OA\Response(
      *  response=204,
      *  description="Delete a message"
      * )
+     *
      * @OA\Tag(name="message")
+     *
      * @Security(name="api_key")
      */
     public function index(
