@@ -35,26 +35,34 @@ class Edit extends ApiController
 
     /**
      * @Route("/messages/{id}", methods={"PUT"})
+     *
      * @OA\RequestBody(
+     *
      *  @OA\Schema(
      *    type="object",
+     *
      *    @OA\Property(property="text", type="string"),
      *    @OA\Property(property="title", type="string"),
      *    @OA\Property(
      *      property="files",
      *      type="array",
+     *
      *      @OA\Items(
      *        type="string",
      *      )
      *    ),
      *  )
      * )
+     *
      * @OA\Response(
      *  response=200,
      *  description="Modify a bookmark",
+     *
      *  @Model(type=App\Entity\Message::class, groups={"read_message"})
      * )
+     *
      * @OA\Tag(name="message")
+     *
      * @Security(name="api_key")
      */
     public function index(
@@ -105,8 +113,8 @@ class Edit extends ApiController
         $this->em->flush();
 
         $message_norm = $this->normalize($message, ['read_message']);
-        $message_norm["preview"] = $this->normalize($message->getPreview(), ['read_message']);
-        $message_norm["author"] = $this->normalize($message->getAuthor(), ['read_message_preview']);
+        $message_norm['preview'] = $this->normalize($message->getPreview(), ['read_message']);
+        $message_norm['author'] = $this->normalize($message->getAuthor(), ['read_message_preview']);
 
         $lineage = [];
         $parent = $message->getParent();
@@ -114,8 +122,7 @@ class Edit extends ApiController
             $lineage[] = $parent->getId();
             $parent = $parent->getParent();
         }
-        $message_norm["lineage"] = $lineage;
-
+        $message_norm['lineage'] = $lineage;
 
         return new JsonResponse(
             $message_norm,
