@@ -2,18 +2,17 @@
 
 namespace App\Controller\Message;
 
-use App\Entity\Message;
 use App\Controller\ApiController;
-use App\Normalizer\ObjectNormalizer;
+use App\Entity\Message;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\Serializer\SerializerInterface;
+use Nelmio\ApiDocBundle\Annotation\Model;
+use Nelmio\ApiDocBundle\Annotation\Security;
+use OpenApi\Annotations as OA;
 use Symfony\Component\ExpressionLanguage\Expression;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use Nelmio\ApiDocBundle\Annotation\Model;
-use Nelmio\ApiDocBundle\Annotation\Security;
-use OpenApi\Annotations as OA;
+use Symfony\Component\Serializer\SerializerInterface;
 
 class Get extends ApiController
 {
@@ -26,12 +25,16 @@ class Get extends ApiController
 
     /**
      * @Route("/messages/{id}", methods={"GET"})
+     *
      * @OA\Response(
      *  response=200,
      *  description="Get a message",
+     *
      *  @Model(type=App\Entity\Message::class, groups={"read_message"})
      * )
+     *
      * @OA\Tag(name="message")
+     *
      * @Security(name="api_key")
      */
     public function index(string $id): Response
@@ -45,8 +48,8 @@ class Get extends ApiController
 
         $this->denyAccessUnlessGranted(new Expression('user in object.getUsersAsArray()'), $message->getGroup());
         $message_norm = $this->normalize($message, ['read_message']);
-        $message_norm["preview"] = $this->normalize($message->getPreview(), ['read_message']);
-        $message_norm["author"] = $this->normalize($message->getAuthor(), ['read_message_preview']);
+        $message_norm['preview'] = $this->normalize($message->getPreview(), ['read_message']);
+        $message_norm['author'] = $this->normalize($message->getAuthor(), ['read_message_preview']);
 
         $lineage = [];
         $parent = $message->getParent();
@@ -54,7 +57,7 @@ class Get extends ApiController
             $lineage[] = $parent->getId();
             $parent = $parent->getParent();
         }
-        $message_norm["lineage"] = $lineage;
+        $message_norm['lineage'] = $lineage;
 
         return new JsonResponse(
             $message_norm,
@@ -64,12 +67,16 @@ class Get extends ApiController
 
     /**
      * @Route("/messages/{id}/preview", methods={"GET"})
+     *
      * @OA\Response(
      *  response=200,
      *  description="Get a message",
+     *
      *  @Model(type=App\Entity\Message::class, groups={"read_message"})
      * )
+     *
      * @OA\Tag(name="message")
+     *
      * @Security(name="api_key")
      */
     public function preview(string $id): Response
@@ -84,7 +91,7 @@ class Get extends ApiController
         $this->denyAccessUnlessGranted(new Expression('user in object.getUsersAsArray()'), $message->getGroup());
 
         $message_preview = $this->normalize($message, ['read_message_preview']);
-        $message_preview["children"] = count($message->getChildren());
+        $message_preview['children'] = count($message->getChildren());
 
         return new JsonResponse($message_preview, JsonResponse::HTTP_OK);
     }
