@@ -5,7 +5,6 @@ namespace App\Controller\User;
 use App\Controller\ApiController;
 use App\Entity\User;
 use Doctrine\ORM\EntityManagerInterface;
-use Nelmio\ApiDocBundle\Annotation\Model;
 use Nelmio\ApiDocBundle\Annotation\Security;
 use OpenApi\Annotations as OA;
 use Symfony\Component\ExpressionLanguage\Expression;
@@ -26,11 +25,14 @@ class Delete extends ApiController
 
     /**
      * @Route("/users/{id}", methods={"DELETE"})
+     *
      * @OA\Response(
      *  response=204,
      *  description="Delete a user",
      * )
+     *
      * @OA\Tag(name="user")
+     *
      * @Security(name="api_key")
      */
     public function index(
