@@ -4,12 +4,12 @@ namespace App\Controller\User;
 
 use App\Controller\ApiController;
 use Doctrine\ORM\EntityManagerInterface;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-use Symfony\Component\Serializer\SerializerInterface;
 use Nelmio\ApiDocBundle\Annotation\Model;
 use Nelmio\ApiDocBundle\Annotation\Security;
 use OpenApi\Annotations as OA;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Serializer\SerializerInterface;
 
 class Me extends ApiController
 {
@@ -22,12 +22,16 @@ class Me extends ApiController
 
     /**
      * @Route("/me", methods={"GET"})
+     *
      * @OA\Response(
      *  response=200,
      *  description="Returns the current logged in user",
+     *
      *  @Model(type=App\Entity\User::class, groups={"read_user", "read_me"})
      * )
+     *
      * @OA\Tag(name="user")
+     *
      * @Security(name="api_key")
      */
     public function index(): Response
