@@ -3,55 +3,67 @@
 namespace App\Entity;
 
 use App\Service\Uuid;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
-use Nelmio\ApiDocBundle\Annotation\Model;
 use OpenApi\Annotations as OA;
 use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
  * @ORM\Table(name="`bookmark`")
+ *
  * @ORM\Entity()
  */
-class Bookmark
+class Bookmark extends ApiEntity
 {
     /**
      * @Assert\NotBlank()
-     * @Groups("*")
+     *
+     * @Groups("public")
+     *
      * @OA\Property(type="guid")
+     *
      * @ORM\Column(type="guid")
+     *
      * @ORM\Id
      */
     private $id;
 
     /**
      * @Assert\NotNull()
+     *
      * @Assert\Type("integer")
-     * @Groups("*")
+     *
+     * @Groups("public")
+     *
      * @OA\Property(type="integer")
+     *
      * @ORM\Column(type="integer")
      */
     private $createdAt;
 
     /**
      * @Assert\NotNull()
-     * @Groups("*")
+     *
+     * @Groups("public")
+     *
      * @OA\Property(type="App\Entity\User")
+     *
      * @ORM\ManyToOne(targetEntity="App\Entity\User", inversedBy="bookmarks")
      */
     private $user;
 
     /**
      * @ORM\ManyToOne(targetEntity="App\Entity\Message", inversedBy="bookmarks")
+     *
      * @Groups({"read_bookmark"})
+     *
      * @OA\Property(type="App\Entity\Message")
      */
     private $message;
 
     /**
-     * @Groups("*")
+     * @Groups("public")
+     *
      * @OA\Property(type="string")
      */
     private $entityType;
