@@ -4,65 +4,81 @@ namespace App\Entity;
 
 use App\Service\Uuid;
 use Doctrine\ORM\Mapping as ORM;
-use Symfony\Component\Validator\Constraints as Assert;
-use Symfony\Component\Serializer\Annotation\Groups;
-use Nelmio\ApiDocBundle\Annotation\Model;
 use OpenApi\Annotations as OA;
+use Symfony\Component\Validator\Constraints as Assert;
 
 /**
  * @ORM\Table(name="`link`")
+ *
  * @ORM\Entity
  */
-class Link
+class Link extends ApiEntity
 {
     /**
      * @ORM\Id
+     *
      * @ORM\Column(type="guid")
+     *
      * @Assert\NotBlank()
+     *
      * @OA\Property(type="guid")
      */
     private $id;
 
     /**
      * @ORM\Column(type="integer")
+     *
      * @Assert\Type("integer")
+     *
      * @Assert\NotNull()
+     *
      * @OA\Property(type="integer")
      */
     private $createdAt;
 
     /**
      * @ORM\Column(type="integer")
+     *
      * @Assert\Type("integer")
+     *
      * @Assert\NotNull()
+     *
      * @OA\Property(type="integer")
      */
     private $updatedAt;
 
     /**
      * @ORM\Column(type="json")
+     *
      * @Assert\NotBlank()
+     *
      * @OA\Property(type="object")
      */
     private $data;
 
     /**
      * @ORM\Column(type="string", unique=true)
+     *
      * @Assert\NotBlank()
+     *
      * @OA\Property(type="string")
      */
     private $url;
 
     /**
      * @ORM\OneToOne(targetEntity="App\Entity\File")
+     *
      * @ORM\JoinColumn(name="preview_id", referencedColumnName="id")
+     *
      * @OA\Property(type="App\Entity\File")
      */
     private $preview;
 
     /**
      * @ORM\Column(type="guid", unique=true)
+     *
      * @Assert\NotBlank()
+     *
      * @OA\Property(type="guid")
      */
     private $secretKey;
