@@ -2,8 +2,8 @@
 
 namespace App\Service;
 
-use App\Entity\Notification as NotificationEntity;
 use App\Entity\Group as GroupEntity;
+use App\Entity\Notification as NotificationEntity;
 use Doctrine\ORM\EntityManagerInterface;
 
 class Group
@@ -19,6 +19,20 @@ class Group
     public function getById($id)
     {
         return $this->em->getRepository(GroupEntity::class)->findOneById($id);
+    }
+
+    public function create($name, $user)
+    {
+        $group = new GroupEntity();
+        $group->setName($name);
+        $this->em->persist($group);
+
+        $user->setLastActivityDate(time());
+        $this->em->persist($user);
+
+        $this->em->flush();
+
+        return $group;
     }
 
     public function addUser($group, $user)
