@@ -4,14 +4,11 @@ namespace App\Service;
 
 use App\Entity\File;
 use App\Entity\Link;
-use App\Service\Image as ImageService;
 use App\Service\Link as LinkService;
 use Doctrine\ORM\EntityManagerInterface;
 use Embed\Embed;
 use GuzzleHttp;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
-use Symfony\Component\Security\Core\Exception\AccessDeniedException;
-use Symfony\Component\Security\Core\Security;
 
 class Url
 {
@@ -48,7 +45,7 @@ class Url
         }
         if (empty($link)) {
             $link = new Link($url);
-            $link->setData(["loading" => true]);
+            $link->setData(['loading' => true]);
         }
 
         // immediatly persist the link to avoid race conditions
@@ -65,22 +62,22 @@ class Url
         $uri = '';
 
         // weak type checks to also accept null until we can add scalar type hints
-        if ($scheme != '') {
-            $uri .= $scheme . ':';
+        if ('' != $scheme) {
+            $uri .= $scheme.':';
         }
 
-        if ($authority != ''|| $scheme === 'file') {
-            $uri .= '//' . $authority;
+        if ('' != $authority || 'file' === $scheme) {
+            $uri .= '//'.$authority;
         }
 
         $uri .= $path;
 
-        if ($query != '') {
-            $uri .= '?' . $query;
+        if ('' != $query) {
+            $uri .= '?'.$query;
         }
 
-        if ($fragment != '') {
-            $uri .= '#' . $fragment;
+        if ('' != $fragment) {
+            $uri .= '#'.$fragment;
         }
 
         return $uri;
@@ -90,8 +87,9 @@ class Url
     public static function exceptionRedirect(string $url): string
     {
         // https://github.com/oscarotero/Embed/issues/458
-        $url = preg_replace("/^https?:\/\/(www.)?youtube.com\/shorts\//", "https://youtube.com/watch?v=", $url);
-        $url = preg_replace("/^https?:\/\/(www.)?youtube.com\/embed\//", "https://youtube.com/watch?v=", $url);
+        $url = preg_replace("/^https?:\/\/(www.)?youtube.com\/shorts\//", 'https://youtube.com/watch?v=', $url);
+        $url = preg_replace("/^https?:\/\/(www.)?youtube.com\/embed\//", 'https://youtube.com/watch?v=', $url);
+
         return $url;
     }
 
@@ -99,24 +97,24 @@ class Url
     {
         try {
             $client = new GuzzleHttp\Client();
-            $res = $client->request('GET', 'https://api.instagram.com/oembed/?url=' . $url);
+            $res = $client->request('GET', 'https://api.instagram.com/oembed/?url='.$url);
             $data = json_decode($res->getBody(), true);
 
             return [
-        'authorName' => $data["author_name"],
-        'authorUrl' => $data["author_url"],
-        'code' => $data["html"],
-        'description' => $data["title"],
-        'thumbnail_url' => $data["thumbnail_url"],
-        'image' => $data["thumbnail_url"],
+        'authorName' => $data['author_name'],
+        'authorUrl' => $data['author_url'],
+        'code' => $data['html'],
+        'description' => $data['title'],
+        'thumbnail_url' => $data['thumbnail_url'],
+        'image' => $data['thumbnail_url'],
         'origin' => $url,
-        'title' => $data["author_name"],
-        'providerName' => $data["provider_name"],
-        'providerUrl' => $data["provider_url"],
+        'title' => $data['author_name'],
+        'providerName' => $data['provider_name'],
+        'providerUrl' => $data['provider_url'],
       ];
         } catch (\Exception $e) {
             return [
-        'origin' => $url, //The original input url
+        'origin' => $url, // The original input url
         'exception' => $e->getMessage(),
       ];
         }
@@ -129,42 +127,42 @@ class Url
             $info = $embed->get(Url::exceptionRedirect($url));
 
             return [
-        'title' => $info->title, //The page title
-        'description' => $info->description, //The page description
-        'url' => Url::composeComponents( //The canonical url
+        'title' => $info->title, // The page title
+        'description' => $info->description, // The page description
+        'url' => Url::composeComponents( // The canonical url
             $info->url->getScheme(),
             $info->url->getAuthority(),
             $info->url->getPath(),
             $info->url->getQuery(),
             $info->url->getFragment(),
         ),
-        'keywords' => $info->keywords, //The page keywords (tags)
+        'keywords' => $info->keywords, // The page keywords (tags)
 
-        'image' => $info->image, //The image choosen as main image
+        'image' => $info->image, // The image choosen as main image
 
-        'code' => $info->code ? $info->code->html : null, //The code to embed the image, video, etc
+        'code' => $info->code ? $info->code->html : null, // The code to embed the image, video, etc
 
-        'authorName' => $info->authorName, //The resource author
-        'authorUrl' => $info->authorUrl, //The author url
+        'authorName' => $info->authorName, // The resource author
+        'authorUrl' => $info->authorUrl, // The author url
 
-        'cms' => $info->cms, //The cms used
-        'language' => $info->language, //The language of the page
-        'languages' => $info->languages, //The alternative languages
+        'cms' => $info->cms, // The cms used
+        'language' => $info->language, // The language of the page
+        'languages' => $info->languages, // The alternative languages
 
-        'providerName' => $info->providerName, //The provider name of the page (Youtube, Twitter, Instagram, etc)
-        'providerUrl' => $info->providerUrl, //The provider url
-        'icon' => $info->icon, //The big icon of the site
-        'favicon' => $info->favicon, //The favicon of the site (an .ico file or a png with up to 32x32px)
+        'providerName' => $info->providerName, // The provider name of the page (Youtube, Twitter, Instagram, etc)
+        'providerUrl' => $info->providerUrl, // The provider url
+        'icon' => $info->icon, // The big icon of the site
+        'favicon' => $info->favicon, // The favicon of the site (an .ico file or a png with up to 32x32px)
 
-        'publishedTime' => $info->publishedTime, //The published time of the resource
-        'license' => $info->license, //The license url of the resource
-        'feeds' => $info->feeds, //The RSS/Atom feeds
-        'content-type' => $info->getResponse()->getHeader('content-Type'), //The content type of the url
-        'origin' => $url, //The original input url
+        'publishedTime' => $info->publishedTime, // The published time of the resource
+        'license' => $info->license, // The license url of the resource
+        'feeds' => $info->feeds, // The RSS/Atom feeds
+        'content-type' => $info->getResponse()->getHeader('content-Type'), // The content type of the url
+        'origin' => $url, // The original input url
       ];
         } catch (\Exception $e) {
             return [
-        'origin' => $url, //The original input url
+        'origin' => $url, // The original input url
         'exception' => $e->getMessage(),
       ];
         }
@@ -173,10 +171,11 @@ class Url
     public static function getData(string $url): array
     {
         $data = Url::getEmbedData($url);
-        if ($data["providerName"] == "Instagram") {
+        if ('Instagram' == $data['providerName']) {
             $instagramData = Url::getInstagramData($url);
             $data = array_merge($data, $instagramData);
         }
+
         return $data;
     }
 }
