@@ -2,7 +2,6 @@ import { h, render } from "preact";
 import App from "./app.js";
 import {
   createBrowserRouter,
-  createRoutesFromElements,
   RouterProvider,
 } from "react-router-dom";
 
@@ -14,7 +13,7 @@ const router = createBrowserRouter(
     { path: "*", Component: App },
   ],
   {
-    basename: `/${document.baseURI.split('/')[3]}`,
+    basename: `/${document.baseURI.split("/")[3]}`,
   }
 );
 
