@@ -7,7 +7,7 @@ import { useState, useEffect, useRef } from "preact/hooks";
 
 // trick from https://stackoverflow.com/a/53837442
 function useForceUpdate() {
-  const [value, setValue] = useState(0); // integer state
+  const setValue = useState(0)[1]; // integer state
   return () => setValue(value => value + 1); // update state to force render
 }
 
